76 SQL test review Add tests to Method(s)
